requirements.txt
blinker==1.4
click==8.0.4
Flask==2.0.3
Flask-Mail==0.9.1
flask-mailman==0.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
mkdocs-material-extensions==1.0.3
python-dotenv==0.19.2
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.0.3
Werkzeug==2.0.3
blinker==1.4
click==8.0.4
Flask==2.0.3
Flask-Mail==0.9.1
flask-mailman==0.3.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
mkdocs-material-extensions==1.0.3
python-dotenv==0.19.2
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.0.3
Werkzeug==2.0.3
